Phytophthora ensemble risk

Calculates Phytophthora cinnamomi's ensemble risk from mechanisms

Usage

phytorisk_ensemble(
  mec_soilwater,
  mec_rootcontact,
  mec_surfacewater,
  mec_zoospread = NULL,
  weights = "equal"
)

Arguments

mec_soilwater

the result of mec_soilwater

mec_rootcontact

the result of mec_rootcontact

mec_surfacewater

the result of mec_surfacewater

mec_zoospread

optional module. The result of mec_zoospread

weights

weights of the ensemble model. The default uses the same weights for each model. The argument accepts a numeric vector with the corresponding weights. See Details

Value

A SpatRaster

Details

This function calculates the ensemble risk as the weighted sum of the individual risk models as follows:

\(Risk = w_1 \cdot Mec_{soilwater} + w_2 \cdot Mec_{rootcontact} + w_3 \cdot Mec_{mec_surfacewater} + w_4 \cdot Mec_{zoospread}\).

Being \(w_i = 0.25\) when mec_zoospread is not null, and \(w_i = \frac{1}{3}\) when mec_zoospread is null. Additionally, the user can specify a numeric vector of length 3 or 4 specifying the desired weights based on expert criteria.

Examples

## TODO