Raw Phytophthora ensemble risk

Calculates Phytophthora cinnamomi's ensemble risk from raw data

Usage

phytorisk_ensemble_raw(
  aoi,
  poi,
  dem,
  treecover,
  weights = "equal",
  th = 100,
  buffer = 50,
  include_zoospread = FALSE,
  ...,
  quiet = FALSE
)

Arguments

aoi

a sf polygon representing the area of interest. Used to mask the tree cover

poi

a single-point sf object denoting the point of interest to run the simulations

dem

a single-band SpatRaster with a digital elevation model

treecover

a single-band SpatRaster where 1 represents host trees, and 0 represents background area

weights

weights of the ensemble model. The default uses the same weights for each model. The argument accepts a numeric vector with the corresponding weights. See phytorisk_ensemble

th

threshold of flow accumulation to delineate streams

buffer

a buffer in meters to extend the spread in every direction

include_zoospread

logical. Whether to include the optional module of mec_zoospread

...

arguments passed to mec_zoospread

quiet

if TRUE, suppress any message or progress bar

Value

A SpatRaster

Details

TO DO